Изменил фронт - настроил ограничение для пользователя в возможности голосования и комментирования в одно действие. Добавил таблицу учёта голосов.

+++ src/main/java/com/javamentor/qa/platform/service/search/conditions/question/SearchQuestionByVotes.java
@@ -21,7 +21,8 @@ public class SearchQuestionByVotes implements Search {
     @Override
     public List<QuestionDto> getList(String searchVotes) {
         if (searchVotes.chars().allMatch(Character::isDigit)) {
-            return searchQuestionDAO.getQuestionsByNumberOfVotes(Integer.parseInt(searchVotes));
+            return searchQuestionDAO.getQuestionsByNumberOfVotes(Long.parseLong(searchVotes));
+//            return searchQuestionDAO.getQuestionsByNumberOfVotes(Integer.parseInt(searchVotes));
         }
         return searchQuestionDAO.getQuestionsSortedByVotes();
     }
